chore(client): separate configuration functions in another file to improve organization
